+++ test/tempo/offset_date_time_test.dart
@@ -256,6 +256,35 @@ void main() {
     });
   });
 
+  group('replace()', () {
+    test('all date fields', () {
+      var dt = OffsetDateTime.withOffset(nepalOffset, 1, 2, 3, 4, 5, 6, 7);
+      var repl = dt.replace(
+          year: 7,
+          month: 6,
+          day: 5,
+          hour: 4,
+          minute: 3,
+          second: 2,
+          nanosecond: 1);
+      expect(repl, hasDateAndTime(7, 6, 5, 4, 3, 2, 1));
+      expect(repl.offset, nepalOffset);
+    });
+
+    test('offset', () {
+      var dt = OffsetDateTime.withOffset(nepalOffset, 1, 2, 3, 4, 5, 6, 7);
+      var repl = dt.replace(offset: nstOffset);
+      expect(repl, hasDateAndTime(1, 2, 3, 4, 5, 6, 7));
+      expect(repl.offset, nstOffset);
+    });
+
+    test('invalid day of month', () {
+      var dt = OffsetDateTime(2025, 12, 31, 4, 3, 2, 1);
+      var repl = dt.replace(month: 2);
+      expect(repl, hasDateAndTime(2025, 2, 28, 4, 3, 2, 1));
+    });
+  });
+
   test('inLeapYear', () {
     expect(OffsetDateTime(1900).inLeapYear, false, reason: 'year = 1900');
     expect(OffsetDateTime(1904).inLeapYear, true, reason: 'year = 1904');